requirements.txt
pandas
scikit-surprise
streamlit